requirements.txt
Django>=5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
redis>=5.0
celery>=5.3
python-dotenv>=1.0
drf-yasg>=1.21
pytest>=7.0
pytest-django>=4.5
pytest-mock>=3.10